ci: add GitHub Actions workflow for the core
Runs clippy (-D warnings), the workspace tests, and the gate scenario suite on
push to main and on PRs. The eBPF gadgets under capture/ are excluded from the
workspace and need a BTF host, so CI does not build them. Also clears a
redundant u128 cast in run.rs that clippy flagged, so deny-warnings is clean.

+++ crates/orchestrate/src/run.rs
@@ -89,7 +89,7 @@ pub fn new_run_id() -> String {
 
     static COUNTER: AtomicU64 = AtomicU64::new(0);
     let now = SystemTime::now().duration_since(UNIX_EPOCH).unwrap_or_default();
-    let ms = (now.as_millis() as u128) & ((1u128 << 48) - 1);
+    let ms = now.as_millis() & ((1u128 << 48) - 1);
 
     let mut h = Sha256::new();
     h.update(now.as_nanos().to_le_bytes());
